@@ -1,33 +1,47 @@
 module github.com/FuzzyMonkeyCo/monkey
 
-go 1.15
+go 1.17
 
 require (
-	github.com/bazelbuild/buildtools v0.0.0-20211129135157-cdedcc0318b9
+	github.com/alecthomas/chroma/v2 v2.0.0-alpha2
+	github.com/bazelbuild/buildtools v0.0.0-20220224152527-5d33f3736e81
+	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
 	github.com/docopt/docopt-go v0.0.0-20180111231733-ee0de3bc6815
 	github.com/fatih/color v1.13.0
-	github.com/getkin/kin-openapi v0.88.0
-	github.com/go-openapi/swag v0.19.15 // indirect
-	github.com/gogo/protobuf v1.3.2
-	github.com/google/gnostic v0.5.7
+	github.com/getkin/kin-openapi v0.90.0
+	github.com/google/gnostic v0.6.6
 	github.com/google/uuid v1.3.0
 	github.com/hashicorp/logutils v1.0.0
-	github.com/mailru/easyjson v0.7.7 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-tty v0.0.3 // indirect
 	github.com/mitchellh/mapstructure v1.4.3
 	github.com/pmezard/go-difflib v1.0.0
 	github.com/stretchr/testify v1.7.0
 	github.com/superhawk610/bar v0.0.2
+	github.com/xeipuuv/gojsonschema v1.2.0
+	go.starlark.net v0.0.0-20220223235035-243c74974e97
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	google.golang.org/grpc v1.44.0
+	google.golang.org/protobuf v1.27.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/dlclark/regexp2 v1.4.0 // indirect
+	github.com/ghodss/yaml v1.0.0 // indirect
+	github.com/go-openapi/jsonpointer v0.19.5 // indirect
+	github.com/go-openapi/swag v0.21.1 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/josharian/intern v1.0.0 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/mattn/go-tty v0.0.4 // indirect
 	github.com/superhawk610/terminal v0.1.0 // indirect
 	github.com/xeipuuv/gojsonpointer v0.0.0-20190905194746-02993c407bfb // indirect
-	github.com/xeipuuv/gojsonschema v1.2.0
-	go.starlark.net v0.0.0-20211203141949-70c0e40ae128
-	golang.org/x/net v0.0.0-20220105145211-5b0dc2dfae98
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e // indirect
+	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415 // indirect
+	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd // indirect
+	golang.org/x/sys v0.0.0-20220224120231-95c6836cb0e7 // indirect
 	golang.org/x/text v0.3.7 // indirect
-	google.golang.org/genproto v0.0.0-20211223182754-3ac035c7e7cb // indirect
-	google.golang.org/grpc v1.43.0
+	google.golang.org/genproto v0.0.0-20220222213610-43724f9ea8cf // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
