@@ -1,104 +1,83 @@
 package openapiv3
 
 import (
-	"fmt"
 	"io"
+
+	"go.starlark.net/starlark"
+	"google.golang.org/protobuf/types/known/structpb"
 
 	"github.com/FuzzyMonkeyCo/monkey/pkg/internal/fm"
 	"github.com/FuzzyMonkeyCo/monkey/pkg/modeler"
-	"github.com/FuzzyMonkeyCo/monkey/pkg/resetter"
-	"github.com/gogo/protobuf/types"
-	"go.starlark.net/starlark"
 )
+
+// Name names the Starlark builtin
+const Name = "openapi3"
+
+// New instanciates a new model
+func New(kwargs []starlark.Tuple) (modeler.Interface, error) {
+	var name, file, host, headerAuthorization starlark.String
+	if err := starlark.UnpackArgs(Name, nil, kwargs,
+		"name", &name,
+		"file", &file,
+		"host??", &host,
+		"header_authorization??", &headerAuthorization,
+	); err != nil {
+		return nil, err
+	}
+	m := &oa3{name: name.GoString()}
+	m.pb = &fm.Clt_Fuzz_Model_OpenAPIv3{
+		File:                file.GoString(),
+		Host:                host.GoString(),
+		HeaderAuthorization: headerAuthorization.GoString(),
+	}
+	return m, nil
+}
 
 var _ modeler.Interface = (*oa3)(nil)
 
-type T = oa3 // TODO: remove, move modelers registry to own package
-
+// oa3 implements a modeler.Interface for use by `monkey`.
 type oa3 struct {
-	fm.Clt_Fuzz_Model_OpenAPIv3
+	name string
 
-	resetter resetter.Interface
+	pb *fm.Clt_Fuzz_Model_OpenAPIv3
 
 	vald *validator
 
 	tcap *tCapHTTP
 }
 
-// ToProto TODO
+// Name uniquely identifies this instance
+func (m *oa3) Name() string { return m.name }
+
+// ToProto marshals a modeler.Interface implementation into a *fm.Clt_Fuzz_Model
 func (m *oa3) ToProto() *fm.Clt_Fuzz_Model {
-	m.Spec = m.vald.Spec
+	m.pb.Spec = m.vald.Spec
 	return &fm.Clt_Fuzz_Model{
-		Model: &fm.Clt_Fuzz_Model_Openapiv3{
-			Openapiv3: &m.Clt_Fuzz_Model_OpenAPIv3,
-		},
+		Name:  m.name,
+		Model: &fm.Clt_Fuzz_Model_Openapiv3{Openapiv3: m.pb},
 	}
 }
 
-// FromProto TODO
-func (m *oa3) FromProto(p *fm.Clt_Fuzz_Model) error {
-	if mm := p.GetOpenapiv3(); mm != nil {
-		m.Clt_Fuzz_Model_OpenAPIv3 = *mm
-		m.vald = &validator{Spec: mm.Spec}
-		return nil
-	}
-	return fmt.Errorf("unexpected model type: %T", p.GetModel())
-}
-
-// SetResetter TODO
-func (m *oa3) SetResetter(sr resetter.Interface) { m.resetter = sr }
-
-// GetResetter TODO
-func (m *oa3) GetResetter() resetter.Interface { return m.resetter }
-
-func (m *oa3) NewFromKwargs(d starlark.StringDict) (modeler.Interface, *modeler.Error) {
-	m = &oa3{}
-	var err *modeler.Error
-
-	if m.File, err = slGetString(d, "file"); err != nil {
-		return nil, err
-	}
-	if m.Host, err = slGetString(d, "host"); err != nil {
-		return nil, err
-	}
-	if m.HeaderAuthorization, err = slGetString(d, "header_authorization"); err != nil {
-		return nil, err
-	}
-
-	return m, nil
-}
-
-func slGetString(d starlark.StringDict, field string) (str string, err *modeler.Error) {
-	var (
-		found bool
-		val   starlark.Value
-	)
-	if val, found = d[field]; found && val.Type() != "string" {
-		err = modeler.NewError(field, "a string", val.Type())
-		return
-	}
-	if found {
-		str = val.(starlark.String).GoString()
-	}
-	return
-}
-
+// InputsCount sums the amount of named schemas or types APIs define
 func (m *oa3) InputsCount() int {
-	return m.vald.InputsCount()
+	return m.vald.inputsCount()
 }
 
+// FilterEndpoints restricts which API endpoints are considered
 func (m *oa3) FilterEndpoints(args []string) ([]eid, error) {
-	return m.vald.FilterEndpoints(args)
+	return m.vald.filterEndpoints(args)
 }
 
-func (m *oa3) Validate(SID sid, data *types.Value) []string {
+func (m *oa3) Validate(SID sid, data *structpb.Value) []string {
 	return m.vald.Validate(SID, data)
 }
 
+// ValidateAgainstSchema tries to smash the data through the given keyhole
 func (m *oa3) ValidateAgainstSchema(absRef string, data []byte) error {
-	return m.vald.ValidateAgainstSchema(absRef, data)
+	return m.vald.validateAgainstSchema(absRef, data)
 }
 
+// WriteAbsoluteReferences pretty-prints the API's named types
 func (m *oa3) WriteAbsoluteReferences(w io.Writer) {
-	m.vald.WriteAbsoluteReferences(w)
+	m.vald.writeAbsoluteReferences(w)
 }
