@@ -3,30 +3,30 @@ package protovalue
 import (
 	"fmt"
 
-	"github.com/gogo/protobuf/types"
+	"google.golang.org/protobuf/types/known/structpb"
 )
 
 // ToGo compiles a google.protobuf.Value value to Go.
 // It panics on unexpected types.
 // See https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.Value
-func ToGo(value *types.Value) interface{} {
+func ToGo(value *structpb.Value) interface{} {
 	switch value.GetKind().(type) {
-	case *types.Value_NullValue:
+	case *structpb.Value_NullValue:
 		return nil
-	case *types.Value_BoolValue:
+	case *structpb.Value_BoolValue:
 		return value.GetBoolValue()
-	case *types.Value_NumberValue:
+	case *structpb.Value_NumberValue:
 		return value.GetNumberValue()
-	case *types.Value_StringValue:
+	case *structpb.Value_StringValue:
 		return value.GetStringValue()
-	case *types.Value_ListValue:
+	case *structpb.Value_ListValue:
 		val := value.GetListValue().GetValues()
 		vs := make([]interface{}, 0, len(val))
 		for _, v := range val {
 			vs = append(vs, ToGo(v))
 		}
 		return vs
-	case *types.Value_StructValue:
+	case *structpb.Value_StructValue:
 		val := value.GetStructValue().GetFields()
 		vs := make(map[string]interface{}, len(val))
 		for n, v := range val {
