@@ -2,7 +2,6 @@ package runtime
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"os"
 	"strings"
@@ -11,8 +10,6 @@ import (
 	"github.com/FuzzyMonkeyCo/monkey/pkg/as"
 	"github.com/FuzzyMonkeyCo/monkey/pkg/internal/fm"
 	"github.com/FuzzyMonkeyCo/monkey/pkg/resetter"
-	"github.com/FuzzyMonkeyCo/monkey/pkg/resetter/shell"
-	"go.starlark.net/starlark"
 )
 
 // Cleanup ensures that resetters are terminated
@@ -25,12 +22,9 @@ func (rt *Runtime) Cleanup(ctx context.Context) (err error) {
 	as.ColorNFO.Println("Cleaning up...")
 
 	log.Println("[NFO] terminating resetter")
-	var rsttr resetter.Interface
-	for _, mdl := range rt.models {
-		rsttr = mdl.GetResetter()
-		break
-	}
-	if errR := rsttr.Terminate(ctx, os.Stdout, os.Stderr); errR != nil && err == nil {
+	if errR := rt.forEachSelectedResetter(ctx, func(name string, rsttr resetter.Interface) error {
+		return rsttr.Terminate(ctx, os.Stdout, os.Stderr)
+	}); errR != nil && err == nil {
 		err = errR
 		// Keep going
 	}
@@ -92,66 +86,20 @@ func (rt *Runtime) reset(ctx context.Context) error {
 }
 
 func (rt *Runtime) runReset(ctx context.Context) (err error) {
-	var rsttr resetter.Interface
-	for _, mdl := range rt.models {
-		rsttr = mdl.GetResetter()
-		break
-	}
-
-	for name, chk := range rt.checks {
-		if err = chk.reset(name); err != nil {
+	if err = rt.forEachCheck(func(name string, chk *check) error {
+		if err := chk.reset(name); err != nil {
 			log.Println("[ERR]", err)
-			return
+			return err
 		}
+		return nil
+	}); err != nil {
+		return
 	}
 	log.Println("[NFO] re-initialized model state")
 
-	stdout := newProgressWriter(rt.progress.Printf)
-	stderr := newProgressWriter(rt.progress.Errorf)
-	err = rsttr.ExecReset(ctx, stdout, stderr, false)
-	return
-}
-
-func newFromKwargs(modelerName string, r starlark.StringDict) (resetter.Interface, error) {
-	const (
-		tExecReset = "ExecReset"
-		tExecStart = "ExecStart"
-		tExecStop  = "ExecStop"
-	)
-	var (
-		ok bool
-		v  starlark.Value
-		vv starlark.String
-		t  string
-		// TODO: other Resetter.s
-		rsttr = &shell.Resetter{}
-	)
-	t = tExecStart
-	if v, ok = r[t]; ok {
-		delete(r, t)
-		if vv, ok = v.(starlark.String); !ok {
-			return nil, fmt.Errorf("%s(%s = ...) must be a string", modelerName, t)
-		}
-		rsttr.Start = vv.GoString()
-	}
-	t = tExecReset
-	if v, ok = r[t]; ok {
-		delete(r, t)
-		if vv, ok = v.(starlark.String); !ok {
-			return nil, fmt.Errorf("%s(%s = ...) must be a string", modelerName, t)
-		}
-		rsttr.Rst = vv.GoString()
-	}
-	t = tExecStop
-	if v, ok = r[t]; ok {
-		delete(r, t)
-		if vv, ok = v.(starlark.String); !ok {
-			return nil, fmt.Errorf("%s(%s = ...) must be a string", modelerName, t)
-		}
-		rsttr.Stop = vv.GoString()
-	}
-	if len(r) != 0 {
-		return nil, fmt.Errorf("unexpected arguments to %s(): %s", modelerName, strings.Join(r.Keys(), ", "))
-	}
-	return rsttr, nil
+	return rt.forEachSelectedResetter(ctx, func(name string, rsttr resetter.Interface) error {
+		stdout := newProgressWriter(rt.progress.Printf)
+		stderr := newProgressWriter(rt.progress.Errorf)
+		return rsttr.ExecReset(ctx, stdout, stderr, false)
+	})
 }
