@@ -155,10 +155,10 @@ func inOrder(t *T, args ...starlark.Value) (starlark.Value, error) {
 	return starlark.None, nil
 }
 
-var errDictOrdering = newInvalidAssertion("values of type dict are not ordered")
+var errDictOrdering = newInvalidAssertion("values of type dict are not considered ordered")
 
 func containsExactlyItemsIn(t *T, args ...starlark.Value) (starlark.Value, error) {
-	arg1 := args[0] // TODO: what when passed **kwargs?
+	arg1 := args[0]
 	if imActual, ok := t.actual.(starlark.IterableMapping); ok {
 		if imExpected, ok := arg1.(starlark.IterableMapping); ok {
 			tt := newT(newTupleSlice(imActual.Items()))
