@@ -131,7 +131,7 @@ func findAttr(name string) (attr, int) {
 
 // LocalThreadKeyForClose is used by Close() and internally to check subjects
 // are eventually resolved.
-var LocalThreadKeyForClose = Default
+var LocalThreadKeyForClose = Module
 
 var εCallFrame = starlark.CallFrame{Pos: syntax.Position{Line: -1, Col: -1}}
 
@@ -142,7 +142,7 @@ var εCallFrame = starlark.CallFrame{Pos: syntax.Position{Line: -1, Col: -1}}
 // test author error.
 func Close(th *starlark.Thread) (err error) {
 	if c, ok := th.Local(LocalThreadKeyForClose).(starlark.CallFrame); ok && c != εCallFrame {
-		err = UnresolvedError(c.Pos.String())
+		err = newUnresolvedError(c.Pos.String())
 	}
 	return
 }
@@ -188,7 +188,7 @@ func builtinAttr(t *T, name string) (starlark.Value, error) {
 			}
 			argz = []starlark.Value{arg1, arg2}
 		default:
-			err := fmt.Errorf("unexpected #args for %s.that(%s).%q(): %d", Default, t.actual.String(), name, nArgs)
+			err := fmt.Errorf("unexpected #args for %s.%s(%s).%q(): %d", Module, Method, t.actual.String(), name, nArgs)
 			return nil, err
 		}
 
