@@ -6,28 +6,32 @@ import (
 	"go.starlark.net/starlark"
 )
 
-// Default is the default module constructor. It is merely the string "assert".
-const Default = "assert"
+var (
+	// Module is the module name used by default.
+	Module = "assert"
+
+	// Method is the attribute name used by default.
+	Method = "that"
+)
 
 type module struct{}
 
-var _ starlark.Value = (*module)(nil)
 var _ starlark.HasAttrs = (*module)(nil)
 
 // NewModule registers a Starlark module of https://truth.dev/
-func NewModule(predeclared starlark.StringDict) { predeclared[Default] = &module{} }
+func NewModule(predeclared starlark.StringDict) { predeclared[Module] = &module{} }
 
-func (m *module) String() string        { return Default }
-func (m *module) Type() string          { return Default }
+func (m *module) String() string        { return Module }
+func (m *module) Type() string          { return Module }
 func (m *module) Freeze()               {}
-func (m *module) Truth() starlark.Bool  { return false }
-func (m *module) Hash() (uint32, error) { return 0, fmt.Errorf("unhashable type: %s", Default) }
-func (m *module) AttrNames() []string   { return []string{"that"} }
+func (m *module) Truth() starlark.Bool  { return true }
+func (m *module) Hash() (uint32, error) { return 0, fmt.Errorf("unhashable type: %s", Module) }
+func (m *module) AttrNames() []string   { return []string{Method} }
 func (m *module) Attr(name string) (starlark.Value, error) {
-	if name != "that" {
+	if name != Method {
 		return nil, nil // no such method
 	}
-	b := starlark.NewBuiltin(Default, That)
+	b := starlark.NewBuiltin(name, That)
 	return b.BindReceiver(m), nil
 }
 
@@ -46,13 +50,12 @@ func That(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwa
 	return newT(target), nil
 }
 
-var _ starlark.Value = (*T)(nil)
 var _ starlark.HasAttrs = (*T)(nil)
 
 func newT(target starlark.Value) *T { return &T{actual: target} }
 
-func (t *T) String() string                           { return fmt.Sprintf("%s.that(%s)", Default, t.actual.String()) }
-func (t *T) Type() string                             { return Default }
+func (t *T) String() string                           { return fmt.Sprintf("%s.%s(%s)", Module, Method, t.actual.String()) }
+func (t *T) Type() string                             { return Module }
 func (t *T) Freeze()                                  { t.actual.Freeze() }
 func (t *T) Truth() starlark.Bool                     { return true }
 func (t *T) Hash() (uint32, error)                    { return 0, fmt.Errorf("unhashable: %s", t.Type()) }
