@@ -13,7 +13,7 @@ func TestIsOrdered(t *testing.T) {
 		s(`(3, 5, 5)`): nil,
 		s(`"abcdef"`):  nil,
 		s(`"fedcba"`):  newTruthAssertion(`Not true that <"fedcba"> is ordered <("f", "e")>.`),
-		s(`{5: 4}`):    newInvalidAssertion("values of type dict are not ordered"),
+		s(`{5: 4}`):    errDictOrdering,
 		s(`(5, 4)`):    newTruthAssertion(`Not true that <(5, 4)> is ordered <(5, 4)>.`),
 		s(`(3, 5, 4)`): newTruthAssertion(`Not true that <(3, 5, 4)> is ordered <(5, 4)>.`),
 	})
@@ -30,7 +30,7 @@ func TestIsOrderedAccordingTo(t *testing.T) {
 		s(`(5, 5, 3)`): nil,
 		s(`"fedcba"`):  nil,
 		s(`"abcdef"`):  newTruthAssertion(`Not true that <"abcdef"> is ordered <("a", "b")>.`),
-		s(`{5: 4}`):    newInvalidAssertion("values of type dict are not ordered"),
+		s(`{5: 4}`):    errDictOrdering,
 		s(`(4, 5)`):    newTruthAssertion(`Not true that <(4, 5)> is ordered <(4, 5)>.`),
 		s(`(3, 5, 4)`): newTruthAssertion(`Not true that <(3, 5, 4)> is ordered <(3, 5)>.`),
 	})
@@ -47,7 +47,7 @@ func TestIsStrictlyOrdered(t *testing.T) {
 		s(`"abcdef"`):  nil,
 		s(`"abcdee"`):  newTruthAssertion(`Not true that <"abcdee"> is strictly ordered <("e", "e")>.`),
 		s(`"fedcba"`):  newTruthAssertion(`Not true that <"fedcba"> is strictly ordered <("f", "e")>.`),
-		s(`{5: 4}`):    newInvalidAssertion("values of type dict are not ordered"),
+		s(`{5: 4}`):    errDictOrdering,
 		s(`(5, 4)`):    newTruthAssertion(`Not true that <(5, 4)> is strictly ordered <(5, 4)>.`),
 		s(`(3, 5, 5)`): newTruthAssertion(`Not true that <(3, 5, 5)> is strictly ordered <(5, 5)>.`),
 	})
@@ -64,7 +64,7 @@ func TestIsStrictlyOrderedAccordingTo(t *testing.T) {
 		s(`"fedcba"`):  nil,
 		s(`"fedcbb"`):  newTruthAssertion(`Not true that <"fedcbb"> is strictly ordered <("b", "b")>.`),
 		s(`"abcdef"`):  newTruthAssertion(`Not true that <"abcdef"> is strictly ordered <("a", "b")>.`),
-		s(`{5: 4}`):    newInvalidAssertion("values of type dict are not ordered"),
+		s(`{5: 4}`):    errDictOrdering,
 		s(`(4, 5)`):    newTruthAssertion(`Not true that <(4, 5)> is strictly ordered <(4, 5)>.`),
 		s(`(5, 5, 3)`): newTruthAssertion(`Not true that <(5, 5, 3)> is strictly ordered <(5, 5)>.`),
 	})
